Move format code into a separate crate
This is a move towards making the runtime distributable without including lots
of complicated parsing code.

We may eventually split binary and text into separate crates as well, so
format-binary can be included with the runtime.

It may also be possible to create a serialized version of the module struct,
so a trivial wire format could be created for wrausmt.

Since the binary parser is in a different crate from the ValueType and
RefType, we have to change from try_from into a custom trait so that it can be
implement. This is a reasonable change regardless.