Simplify BinaryParser code structure by making it a struct (#221)
The existing everything-is-a-trait implementation is cool, but it was getting
unwieldy. I ran into a puzzle when trying to add a Tracer, when I tried to
pass the Tracer along with the limited "Take" readers. There may have been a
solution to the problem, but overall the code was very difficult to reason
about, with little gain.

If I ever start a rust-explorations project, the trait-based puzzle I
encountered will definitely be featured.