Propertly implement blocktypes (#311)
There are a few "special" blocktypes, the void type and the single value
result types. These don't correspond to an indexed type, instead the emitter
and runtime just handles them directly.

This also changes to emitting the function arity rather than the index, since
that's what's needed at runtime, and it makes it easier to handle these
special non-index types.

Also remove the "void" value, which doesn't really exist in WASM, and was just
a bad workaround for trying to populate an error parameter.